# Conflicts:
#	app/src/main/java/com/nightstalker/artic/core/data/model/artwork/detail/ArtworkData.kt
#	app/src/main/java/com/nightstalker/artic/features/MappersExt.kt
#	app/src/main/java/com/nightstalker/artic/features/artwork/presentation/ui/detail/ArtworkDetailsFragment.kt
#	app/src/main/java/com/nightstalker/artic/features/exhibition/presentation/ui/detail/ExhibitionDetailsFragment.kt
#	app/src/main/res/layout/fragment_exhibition_details.xml
#	app/src/main/res/navigation/nav_main.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/nightstalker/artic/features/artwork/presentation/ui/list/ArtworksListAdapter.kt
@@ -37,7 +37,7 @@ class ArtworksListAdapter(
             Glide.with(context).load(imageUrl).into(placeImage)
 
             root.setOnClickListener {
-                onItemClicked(item.id)
+                onItemClicked(item.id.toInt())
             }
         }
     }
